requirements.txt
altair==5.2.0
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.0
Counter==1.0.0
cryptography==42.0.5
cycler==0.12.1
docx2txt==0.8
fonttools==4.50.0
gitdb==4.0.11
GitPython==3.1.42
idna==3.6
Jinja2==3.1.3
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
lxml==5.1.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.3
mdurl==0.1.2
nltk==3.8.1
numpy==1.26.4
packaging==23.2
pandas==2.2.1
pdfminer.six==20231228
pdfplumber==0.11.0
pillow==10.2.0
protobuf==4.25.3
pyarrow==15.0.1
pycparser==2.21
pydeck==0.8.1b0
Pygments==2.17.2
pyparsing==3.1.2
PyPDF2==3.0.1
pypdfium2==4.28.0
python-dateutil==2.9.0.post0
python-docx==1.1.0
pytz==2024.1
referencing==0.34.0
regex==2023.12.25
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
six==1.16.0
smmap==5.0.1
soupsieve==2.5
streamlit==1.32.2
tenacity==8.2.3
toml==0.10.2
toolz==0.12.1
tornado==6.4
tqdm==4.66.2
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
watchdog==4.0.0
wordcloud==1.9.3
unidecode
bertopic
umap-learn
textblob
spacy
wordcloud
pyLDAvis==2.1.2
ipykernel
git+https://github.com/scikit-learn-contrib/hdbscan.git
google-play-scraper
streamlit
